RAG/requirements.txt
# Legal RAG — retrieval + generation + safety audit (Python 3.10+)
# vLLM / torch: see https://docs.vllm.ai for CUDA notes

# --- Core RAG ---
sentence-transformers>=3.0.0
torch>=2.1.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
pymupdf>=1.24.0
python-docx>=1.1.0
vllm>=0.6.3
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.66.0

# --- Safety audit (hexphi_generate_audit, harmful_score / Llama Guard, ASR) ---
datasets>=2.14.0
transformers>=4.40.0
openai>=1.40.0
